+++ CHANGELOG.md
@@ -16,6 +16,10 @@ Exceptions are acceptable depending on the circumstances (critical bug fixes tha
 
 ## [Unreleased]
 
+### Fixed
+
+- fixed the bug creating empty PRs with just the `CHANGELOG.md` changed
+
 ## [0.2.1] - 2026-02-09
 
 ### Changed